8305842: Replace qtdemux with MP4 demux based on Media Foundation [v9]

+++ modules/javafx.media/src/main/native/gstreamer/plugins/progressbuffer/win32/filecache.c
@@ -167,7 +167,7 @@ GstFlowReturn cache_read_buffer_from_position2(Cache* cache, gint64 start_positi
         return GST_FLOW_FLUSHING;
 
     if ((guint64)size > bytes_available)
-        size = bytes_available;
+        size = (guint)bytes_available;
 
     return cache_read_buffer_from_position(cache, start_position, size, buffer);
 }
